+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### BC breaks
 
+* raised minimum required PHP version to 7.3
 * `stubbles\input\broker\RequestBroker::procure()` doesn't return the given and procured object instance any more
 * `stubbles\input\filter\range\StringLength` doesn't support instances of `stubbles\values\Secret` any more
 * `stubbles\input\ValueReader::asSecret()` doesn't take a `stubbles\input\filter\range\StringLength` any more but `stubbles\input\filter\range\SecretMinLength`